+++ react-frontend/src/pages/admin/ViewDetailClass.jsx
@@ -1,11 +1,8 @@
 import React from "react";
-import ava from "../assets/images/avatar-default.jpg";
-import { useNavigate } from "react-router-dom";
 import SidebarforAdmin from "./SidebarforAdmin";
 import HeaderAdmin from "./HeaderAdmin";
 
-function ViewDetailsUser() {
-  const navigate = useNavigate();
+function ViewDetailClass() {
   return (
     <div className="container-fluid">
       <div className="row">
@@ -13,43 +10,32 @@ function ViewDetailsUser() {
         <div className="col-sm">
           <HeaderAdmin />
           <div className="card mb-4">
-            <div className="card-header fs-5 fw-bold text-uppercase">Account Details</div>
+            <div className="card-header fs-5 fw-bold text-uppercase">
+              Class Details
+            </div>
             <div className="card-body">
               <form>
                 <div className="mb-3">
-                  <label className="small mb-1 fs-6">Profile Picture </label>
-                  <img
-                    className="img-account-profile rounded-circle"
-                    src={ava}
-                    alt=""
-                    style={{
-                      width: "100px",
-                      height: "100px",
-                      marginLeft: "35%",
-                    }}
-                  />
-                </div>
-                <div className="mb-3">
-                  <label className="small mb-1 fs-6">Username </label>
+                  <label className="small mb-1 fs-6">Class Name </label>
                   <input
                     className="form-control"
                     type="text"
                     readOnly
-                    value="username"
+                    value="N2"
                   />
                 </div>
                 <div className="row gx-3 mb-3">
                   <div className="col-md-6">
-                    <label className="small mb-1 fs-6">First name</label>
+                    <label className="small mb-1 fs-6">Class ID</label>
                     <input
                       className="form-control"
                       type="text"
                       readOnly
-                      value="ABC"
+                      value="12"
                     />
                   </div>
                   <div className="col-md-6">
-                    <label className="small mb-1 fs-6">Last name</label>
+                    <label className="small mb-1 fs-6">Tutor guide</label>
                     <input
                       className="form-control"
                       type="text"
@@ -61,7 +47,7 @@ function ViewDetailsUser() {
 
                 <div className="row gx-3 mb-3">
                   <div className="col-md-6">
-                    <label className="small mb-1 fs-6">Date of birth</label>
+                    <label className="small mb-1 fs-6">Create Date</label>
                     <input
                       className="form-control"
                       type="text"
@@ -71,56 +57,34 @@ function ViewDetailsUser() {
                   </div>
 
                   <div className="col-md-6">
-                    <label className="small mb-1 fs-6">Address</label>
+                    <label className="small mb-1 fs-6">Member's Joined</label>
                     <input
                       className="form-control"
                       readOnly
                       type="text"
-                      value="San Francisco, CA"
+                      value="50"
                     />
                   </div>
                 </div>
 
                 <div className="row gx-3 mb-3">
                   <div className="col-md-6">
-                    <label className="small mb-1 fs-6">Email address</label>
+                    <label className="small mb-1 fs-6">Description</label>
                     <input
                       className="form-control"
                       type="tel"
                       readOnly
-                      value="name@example.com"
+                      value="Hanh trinh di den N2"
                     />
                   </div>
 
-                  <div className="col-md-6">
-                    <label className="small mb-1 fs-6">Phone number</label>
-                    <input
-                      className="form-control"
-                      type="tel"
-                      readOnly
-                      value="0123456789"
-                    />
-                  </div>
-                </div>
-
-                <div className="row gx-3 mb-3">
                   <div className="col-md-6">
                     <label className="small mb-1 fs-6">Status</label>
                     <input
                       className="form-control"
                       type="tel"
                       readOnly
-                      value="Pending"
-                    />
-                  </div>
-
-                  <div className="col-md-6">
-                    <label className="small mb-1 fs-6">Role</label>
-                    <input
-                      className="form-control"
-                      type="text"
-                      readOnly
-                      value="Learner"
+                      value="Active"
                     />
                   </div>
                 </div>
@@ -138,4 +102,4 @@ function ViewDetailsUser() {
   );
 }
 
-export default ViewDetailsUser;
+export default ViewDetailClass;
