+++ react-frontend/src/pages/admin/ViewDetailsUser.jsx
@@ -1,8 +1,11 @@
-import React from 'react'
+import React from "react";
+import ava from "../../assets/images/avatar-default.jpg";
+import { useNavigate } from "react-router-dom";
 import SidebarforAdmin from "./SidebarforAdmin";
 import HeaderAdmin from "./HeaderAdmin";
 
-function ViewDetailClass() {
+function ViewDetailsUser() {
+  const navigate = useNavigate();
   return (
     <div className="container-fluid">
       <div className="row">
@@ -10,30 +13,45 @@ function ViewDetailClass() {
         <div className="col-sm">
           <HeaderAdmin />
           <div className="card mb-4">
-            <div className="card-header fs-5 fw-bold text-uppercase">Class Details</div>
+            <div className="card-header fs-5 fw-bold text-uppercase">
+              Account Details
+            </div>
             <div className="card-body">
               <form>
                 <div className="mb-3">
-                  <label className="small mb-1 fs-6">Class Name </label>
+                  <label className="small mb-1 fs-6">Profile Picture </label>
+                  <img
+                    className="img-account-profile rounded-circle"
+                    src={ava}
+                    alt=""
+                    style={{
+                      width: "100px",
+                      height: "100px",
+                      marginLeft: "35%",
+                    }}
+                  />
+                </div>
+                <div className="mb-3">
+                  <label className="small mb-1 fs-6">Username </label>
                   <input
                     className="form-control"
                     type="text"
                     readOnly
-                    value="N2"
+                    value="username"
                   />
                 </div>
                 <div className="row gx-3 mb-3">
                   <div className="col-md-6">
-                    <label className="small mb-1 fs-6">Class ID</label>
+                    <label className="small mb-1 fs-6">First name</label>
                     <input
                       className="form-control"
                       type="text"
                       readOnly
-                      value="12"
+                      value="ABC"
                     />
                   </div>
                   <div className="col-md-6">
-                    <label className="small mb-1 fs-6">Tutor guide</label>
+                    <label className="small mb-1 fs-6">Last name</label>
                     <input
                       className="form-control"
                       type="text"
@@ -45,7 +63,7 @@ function ViewDetailClass() {
 
                 <div className="row gx-3 mb-3">
                   <div className="col-md-6">
-                    <label className="small mb-1 fs-6">Create Date</label>
+                    <label className="small mb-1 fs-6">Date of birth</label>
                     <input
                       className="form-control"
                       type="text"
@@ -55,34 +73,56 @@ function ViewDetailClass() {
                   </div>
 
                   <div className="col-md-6">
-                    <label className="small mb-1 fs-6">Member's Joined</label>
+                    <label className="small mb-1 fs-6">Address</label>
                     <input
                       className="form-control"
                       readOnly
                       type="text"
-                      value="50"
+                      value="San Francisco, CA"
                     />
                   </div>
                 </div>
 
                 <div className="row gx-3 mb-3">
                   <div className="col-md-6">
-                    <label className="small mb-1 fs-6">Description</label>
+                    <label className="small mb-1 fs-6">Email address</label>
                     <input
                       className="form-control"
                       type="tel"
                       readOnly
-                      value="Hanh trinh di den N2"
+                      value="name@example.com"
                     />
                   </div>
 
+                  <div className="col-md-6">
+                    <label className="small mb-1 fs-6">Phone number</label>
+                    <input
+                      className="form-control"
+                      type="tel"
+                      readOnly
+                      value="0123456789"
+                    />
+                  </div>
+                </div>
+
+                <div className="row gx-3 mb-3">
                   <div className="col-md-6">
                     <label className="small mb-1 fs-6">Status</label>
                     <input
                       className="form-control"
                       type="tel"
                       readOnly
-                      value="Active"
+                      value="Pending"
+                    />
+                  </div>
+
+                  <div className="col-md-6">
+                    <label className="small mb-1 fs-6">Role</label>
+                    <input
+                      className="form-control"
+                      type="text"
+                      readOnly
+                      value="Learner"
                     />
                   </div>
                 </div>
@@ -97,7 +137,7 @@ function ViewDetailClass() {
         </div>
       </div>
     </div>
-  )
+  );
 }
 
-export default ViewDetailClass
+export default ViewDetailsUser;
